Reference

All docstrings!

CommonLicenses.CommonLicenses — Module

Display open-source software (and other) licenses inline in your executable documents!

Extended help

README

CommonLicenses.jl

Executable licenses for executable documents!

Installation

Choose one of the lines below!

import Pkg; Pkg.add("CommonLicenses")

# In a Julia REPL
]add CommonLicenses

Motivation

Why does this exist?

Imagine you're writing a blog post. Or a Pluto notebook. Or a Jupyter notebook. Or whatever! You're writing words alongside content, and you'd like to license your work. The license needs to stand alone with your document; the "naive" solution is to find the contents of your standard license file, and paste the license into a block of text somewhere in your document. Cool. Two weeks later you're writing another blog post, or whatever. You need to re-find the contents of this license. Not a big deal, but this gets a bit cumbersome, right?

Enter: CommonLicenses.jl. This package provides every standard license tracked by the SPDX License List, so all you need to do is install this package, and print your desired license wherever you want it!

Usage

Try it out for yourself!

julia> using CommonLicenses

julia> using CommonLicenses: MIT, Unlicense

julia> MIT()

julia> Unlicense()

julia> CommonLicenses.text(MIT(copyright="Joe(y)"))

julia> CommonLicenses.name(License("CC-BY-4.0"))

Credits

Built on top of lots of helpful projects!

This package was developed independently of PkgLicenses.jl, though it does look like JuliaPackaging got there first! This package adds goodies on top of the functionality of PkgLicenses.jl, including MIME type support, and licenses for non-software projects such as Creative Commons.

The SPDX License List is used to fetch the contents for each requested license. The GitHub License API is used to find additional metadata for each license, if available.

License

MIT License

Copyright (c) 2023 Joe Carpinelli

Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

Exports

License

Imports

Base

Core

Dates

DocStringExtensions

HTTP

JSON

Scratch

CommonLicenses.AbstractLicense — Type

abstract type AbstractLicense

A supertype for all licenses.

Extended Help

All License subtypes should have the following methods defined:

CommonLicenses.name(license)::AbstractString

CommonLicenses.spdx(license)::AbstractString

CommonLicenses.text(license)::AbstractString

The following methods may be optionally defined.

CommonLicenses.conditions(license)::AbstractSet{<:AbstractString}

CommonLicenses.permissions(license)::AbstractSet{<:AbstractString}

CommonLicenses.limitations(license)::AbstractSet{<:AbstractString}

CommonLicenses.arguments(license)::NTuple{N,Pair{<:AbstractString,Symbol}} where {N}

CommonLicenses.defaults(license)::NTuple{N,Pair{Symbol,<:Any}} where {N}

CommonLicenses.License — Type

struct License{ID, C<:(Union{Missing, var"#s12"} where var"#s12"<:(AbstractSet{<:AbstractString})), P<:(Union{Missing, var"#s14"} where var"#s14"<:(AbstractSet{<:AbstractString})), L<:(Union{Missing, var"#s16"} where var"#s16"<:(AbstractSet{<:AbstractString}))} <: CommonLicenses.AbstractLicense

Metadata for a legal license. When an instance of this type is Base.shown, the content is displayed in a software-license-typical format: i.e. as you would see in a software repository's LICENSE file.

CommonLicenses.License — Method

Construct a license using the SPDX identifier, and any modifications you wish to make to the license text.

CommonLicenses.License — Method

Construct a license, using keyword arguments to replace the keywords specified by arguments(License{ID}).

CommonLicenses.arguments — Function

Specify keywords to replace within the license's text.

Extended Help

Some licenses include small bits of custom information, i.e. the date and name at the top of many MIT licenses. Defining this method for a license type allows a user to specify this information in the argument of the license constructor. If no specific keywords are provided, a fallback method returns an empty NamedTuple.

CommonLicenses.conditions — Function

Return a hint at the conditions of the license. This is provided through GitHub's License API, which itself depends on the licensee Ruby project.

Extended Help

References

GitHub License Endpoint: https://api.github.com/licenses

GitHub License API Docs: https://docs.github.com/en/rest/licenses#get-a-license

GitHub Page for licensee: https://github.com/licensee/licensee

CommonLicenses.defaults — Function

Specify default values for each license argument.

Extended Help

See also, CommonLicenses.arguments.

CommonLicenses.limitations — Function

Return a hint at the limitations of the license. This is provided through GitHub's License API, which itself depends on the licensee Ruby project.

Extended Help

References

GitHub License Endpoint: https://api.github.com/licenses

GitHub License API Docs: https://docs.github.com/en/rest/licenses#get-a-license

GitHub Page for licensee: https://github.com/licensee/licensee

CommonLicenses.name — Function

Return the official name for the license.

CommonLicenses.permissions — Function

Return a hint at the permissions of the license. This is provided through GitHub's License API, which itself depends on the licensee Ruby project.

Extended Help

References

GitHub License Endpoint: https://api.github.com/licenses

GitHub License API Docs: https://docs.github.com/en/rest/licenses#get-a-license

GitHub Page for licensee: https://github.com/licensee/licensee

CommonLicenses.required — Method

required(license)

Return all required arguments.

Extended Help

See also, CommonLicenses.arguments.

CommonLicenses.spdx — Function

Return the SPDX identifier for the license.

CommonLicenses.text — Function

Return the text of the license, with replacement arguments specified if necessary.